Added method to api type to check if apprenticeship is in a status to be approved.
