+++ src/SFA.DAS.Commitments.Api.Types.UnitTests/SFA.DAS.Commitments.Api.Types.UnitTests.csproj
@@ -52,6 +52,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="WhenApprenticeshipStatus.cs" />
     <Compile Include="WhenCommitmentListItemStatus.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
