+++ src/SFA.DAS.Commitments.Api.Types/Apprenticeship.cs
@@ -17,5 +17,10 @@ namespace SFA.DAS.Commitments.Api.Types
         public AgreementStatus AgreementStatus { get; set; }
 
         public string ApprenticeshipName => $"{FirstName} {LastName}";
+
+        public bool CanBeApproved()
+        {
+            return Status == ApprenticeshipStatus.ReadyForApproval;
+        }
     }
 }
